T -0 day work up to 11:30

+++ TEAM2 ONE MONTH CODE/src/Drivers/Outputs/LED/LED_driver.c
@@ -7,7 +7,7 @@
   /* LED method for two XPEBGR-L1-0000-00G01-SB01 LEDs */
   #include <asf.h>
   #include "Drivers/Outputs/LED/LED_driver.h"
-  void LED(uint32_t period,uint8_t duty_cycle){
+  void LED(uint16_t period,uint8_t duty_cycle){
 	  PORTD.OUT |= 0b00000010; //out, high = 1, low = 0.
 
 	  TCD0.PER = period; //Set period of output.
